requirements.txt
alembic==1.8.0; python_version >= "3.7"
attrs==21.4.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
autopage==0.5.1; python_version >= "3.6"
beautifulsoup4==4.11.1; python_full_version >= "3.6.0"
cachetools==5.0.0; python_version >= "3.7" and python_version < "4.0" and (python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "3.11" or python_version >= "3.6" and python_version < "3.11" and python_full_version >= "3.6.0")
certifi==2021.10.8; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "3.11" or python_full_version >= "3.6.0" and python_version >= "3.6" and python_version < "3.11"
charset-normalizer==2.0.12; python_full_version >= "3.6.0" and python_version >= "3.6" and python_version < "3.11"
click==8.0.4; python_version >= "3.6"
cliff==3.10.1; python_version >= "3.6"
cmaes==0.8.2; python_version >= "3.6"
cmd2==2.4.1; python_version >= "3.6"
colorama==0.4.4; python_version >= "3.6" and python_full_version < "3.0.0" and sys_platform == "win32" and platform_system == "Windows" or sys_platform == "win32" and python_version >= "3.6" and python_full_version >= "3.5.0" and platform_system == "Windows"
colorlog==6.6.0; python_version >= "3.6"
db-dtypes==1.0.0; python_version >= "3.6" and python_version < "3.11"
google-api-core==2.7.2; python_version >= "3.6" and python_version < "3.11"
google-auth==2.6.6; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "3.11" or python_version >= "3.6" and python_version < "3.11" and python_full_version >= "3.6.0"
google-cloud-bigquery-storage==2.13.1; python_version >= "3.6" and python_version < "3.11"
google-cloud-bigquery==3.0.1; python_version >= "3.6" and python_version < "3.11"
google-cloud-core==2.3.0; python_version >= "3.6" and python_version < "3.11"
google-crc32c==1.3.0; python_version >= "3.6" and python_version < "3.11"
google-resumable-media==2.3.2; python_version >= "3.6" and python_version < "3.11"
googleapis-common-protos==1.56.0; python_version >= "3.6" and python_version < "3.11"
greenlet==1.1.2; python_version >= "3" and python_full_version < "3.0.0" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6") and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7") or python_version >= "3" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6") and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7") and python_full_version >= "3.5.0"
grpcio-status==1.45.0; python_version >= "3.6" and python_version < "3.11"
grpcio==1.45.0; python_version >= "3.6" and python_version < "3.11"
idna==3.3; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "3.11" or python_full_version >= "3.6.0" and python_version >= "3.6" and python_version < "3.11"
loguru==0.6.0; python_version >= "3.5"
mako==1.2.0; python_version >= "3.7"
markupsafe==2.1.1; python_version >= "3.7"
numpy==1.22.3
optuna==2.10.0; python_version >= "3.6"
packaging==21.3; python_version >= "3.6" and python_version < "3.11"
pandas==1.4.2; python_version >= "3.8"
pbr==5.9.0; python_version >= "3.6"
plotly==5.7.0; python_version >= "3.6"
prettytable==3.3.0; python_version >= "3.7"
proto-plus==1.20.3; python_version >= "3.6" and python_version < "3.11"
protobuf==3.20.0; python_version >= "3.7" and python_version < "3.11"
pyarrow==7.0.0; python_version >= "3.7" and python_version < "3.11"
pyasn1-modules==0.2.8; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "3.11" or python_version >= "3.6" and python_version < "3.11" and python_full_version >= "3.6.0"
pyasn1==0.4.8; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "3.11" and (python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "3.11" or python_version >= "3.6" and python_version < "3.11" and python_full_version >= "3.6.0") or python_version >= "3.6" and python_version < "3.11" and python_full_version >= "3.6.0" and (python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "3.11" or python_version >= "3.6" and python_version < "3.11" and python_full_version >= "3.6.0")
pybit==2.1.0; python_version >= "3.6"
pydantic==1.9.0; python_full_version >= "3.6.1"
pyparsing==3.0.7; python_version >= "3.6" and python_version < "3.11"
pyperclip==1.8.2; python_version >= "3.6"
pyreadline3==3.4.1; sys_platform == "win32" and python_version >= "3.6"
python-dateutil==2.8.2; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.3.0")
python-dotenv==0.19.2; python_version >= "3.5"
pytz==2022.1; python_version >= "3.8" and python_version < "3.11"
pyyaml==6.0; python_version >= "3.6"
requests==2.27.1; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "3.11" or python_full_version >= "3.6.0" and python_version >= "3.6" and python_version < "3.11"
rsa==4.8; python_version >= "3.6" and python_version < "4" and (python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "3.11" or python_version >= "3.6" and python_version < "3.11" and python_full_version >= "3.6.0")
scipy==1.8.1; python_version >= "3.8" and python_version < "3.11"
six==1.16.0; python_version >= "3.8" and python_full_version < "3.0.0" and python_version < "3.11" or python_full_version >= "3.6.0" and python_version >= "3.8" and python_version < "3.11"
soupsieve==2.3.2.post1; python_version >= "3.6" and python_full_version >= "3.6.0"
sqlalchemy==1.4.37; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7"
stevedore==3.5.0; python_version >= "3.6"
tenacity==8.0.1; python_version >= "3.6"
tqdm==4.64.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
typer==0.4.1; python_version >= "3.6"
typing-extensions==4.1.1; python_version >= "3.6" and python_full_version >= "3.6.1"
urllib3==1.26.9; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "3.11" or python_full_version >= "3.6.0" and python_version < "3.11" and python_version >= "3.6"
wcwidth==0.2.5; python_version >= "3.7"
websocket-client==1.3.2; python_version >= "3.7"
websockets==10.2; python_version >= "3.7"
win32-setctime==1.1.0; sys_platform == "win32" and python_version >= "3.5"
